+++ ScavgameTranslationUtils/ViewModels/TranslationWindowViewModel.cs
@@ -11,131 +11,199 @@ namespace ScavgameTranslationUtils.ViewModels;
 
 public partial class TranslationWindowViewModel : ObservableObject
 {
-    public List<int> AllowedIndents { get; } = [2, 3, 4];
-
     public partial class DisplayNode(DisplayNode? parent) : ObservableObject
     {
-        public DisplayNode? Parent { get; }= parent;
+        public DisplayNode? Parent { get; } = parent;
         
-        public string DisplayLabel => AllItemsTranslated ? Label : $"{Label} (*)";
+        public string PathPartView => AllItemsTranslated ? PathPartDisplay : $"{PathPartDisplay} (*)";
+        public string FullPathView => AllItemsTranslated ? FullPathDisplay : $"{FullPathDisplay} (*)";
         
-        [ObservableProperty]
-        [NotifyPropertyChangedFor(nameof(DisplayLabel))]
-        private string _label = "";
+        public required string PathPart { get; init; }
+        public required string FullPath { get; init; }
+        
+        public required string PathPartDisplay { get; init; }
+        public required string FullPathDisplay { get; init; }
 
-        [ObservableProperty]
-        private ObservableCollection<DisplayNode> _subNodes = [];
+        public ObservableCollection<DisplayNode> SubNodes { get; } = [];
 
-        [ObservableProperty]
-        private string? _terminalKey;
+        public bool IsTerminalNode => SubNodes.Count == 0;
 
-        [NotifyPropertyChangedFor(nameof(DisplayLabel))]
+        [NotifyPropertyChangedFor(nameof(PathPartView))]
+        [NotifyPropertyChangedFor(nameof(FullPathView))]
         [ObservableProperty]
         private bool _allItemsTranslated;
+
+        // This is a shallow check
+        public void UpdateTranslationStatus(Workspace workspace)
+        {
+            if (IsTerminalNode)
+            {
+                AllItemsTranslated = workspace.GetText(FullPath) != null;
+            }
+            else
+            {
+                bool allTranslated = true;
+
+                foreach (var node in SubNodes)
+                    allTranslated = allTranslated && node.AllItemsTranslated;
+                
+
+                AllItemsTranslated = allTranslated;
+            }
+            
+            Parent?.UpdateTranslationStatus(workspace);
+        }
+
+        public override string ToString() => $"FullPath={FullPath}, IsTranslated={AllItemsTranslated}";
     }
     
-    public partial class DisplayKey(Workspace workspace) : ObservableObject
+    public class NavigationData
     {
-        public string Display => IsTranslated 
-            ? Constants.TurnToDisplayPathJoined(workspace, Key) 
-            : $"{Constants.TurnToDisplayPathJoined(workspace, Key)} (*)";
+        public static NavigationData Create(Workspace workspace, bool sortInnerKeysAlphabetically)
+        {
+            var treeViewNodes = new ObservableCollection<DisplayNode>();
+
+            foreach (var path in workspace.Paths)
+            {
+                ObservableCollection<DisplayNode> parentList = treeViewNodes;
+                DisplayNode? parentNode = null;
+
+                var (displayParts, parts) = Constants.SplitToDisplayAndParts(workspace, path);
+
+                for (int i = 0; i < parts.Length; i++)
+                {
+                    DisplayNode? targetNode = parentList.FirstOrDefault(x => x.PathPart == parts[i]);
+
+                    if (targetNode == null)
+                    {
+                        parentList.Add(targetNode = new DisplayNode(parentNode)
+                        {
+                            PathPart = parts[i],
+                            FullPath = parentNode != null ? $"{parentNode.FullPath}:{parts[i]}" : parts[i],
+                            PathPartDisplay = displayParts[i],
+                            FullPathDisplay = parentNode != null ? $"{parentNode.FullPathDisplay} > {displayParts[i]}" : displayParts[i]
+                        });
+                    }
+
+                    parentList = targetNode.SubNodes;
+                    parentNode = targetNode;
+                }
+            }
+            
+            if (sortInnerKeysAlphabetically)
+            {
+                // TODO: Bad string comparison here
+                List<DisplayNode?> nodesToSort = new[] { "main", "buildings", "moodles", "other" }
+                    .Select(label => treeViewNodes.FirstOrDefault(node => node.PathPart.Equals(label, StringComparison.InvariantCultureIgnoreCase)))
+                    .ToList();
+
+                var characters = treeViewNodes.FirstOrDefault(node => node.PathPart.Equals("character", StringComparison.InvariantCultureIgnoreCase));
+
+                if (characters != null)
+                    nodesToSort.AddRange(characters.SubNodes);
+
+                foreach (var node in nodesToSort.Where(x => x != null).Cast<DisplayNode>())
+                {
+                    var sortedNodes = node.SubNodes.OrderBy(x => x.PathPart).ToArray();
+                    node.SubNodes.Clear();
+                    
+                    foreach (var sortedNode in sortedNodes)
+                        node.SubNodes.Add(sortedNode);
+                }
+            }
+
+            var listViewNodes = new ObservableCollection<DisplayNode>();
+            
+            var stack = new Stack<DisplayNode>(treeViewNodes.Reverse());
+
+            while (stack.Count > 0)
+            {
+                var nextNode = stack.Pop();
+
+                if (nextNode.IsTerminalNode)
+                {
+                    listViewNodes.Add(nextNode);
+                }
+                else
+                {
+                    foreach (var node in nextNode.SubNodes.Reverse())
+                        stack.Push(node);
+                }
+            }
+
+            var categoryStarts = new List<(string Category, int Start)>();
+            
+            for (int i = 0; i < listViewNodes.Count; i++)
+            {
+                var category = GetKeyCategory(listViewNodes[i].FullPath);
+                if (categoryStarts.FindIndex(x => x.Category == category) == -1)
+                    categoryStarts.Add((category, i));
+            }
+
+            return new NavigationData()
+            {
+                CategoryStarts = categoryStarts,
+                TreeView = treeViewNodes,
+                ListView = listViewNodes,
+            };
+        }
         
-        [ObservableProperty]
-        [NotifyPropertyChangedFor(nameof(Display))]
-        private string _key = "";
-
-        [ObservableProperty]
-        [NotifyPropertyChangedFor(nameof(Display))]
-        private bool _isTranslated;
+        public required ObservableCollection<DisplayNode> TreeView { get; init; }
+        public required ObservableCollection<DisplayNode> ListView { get; init; }
+        public required List<(string Category, int Start)> CategoryStarts { get; init; }
     }
-
-    private List<(string Category, int Start)> CategoryStarts { get; } = [];
     
     public TranslationWindowViewModel(Workspace workspace)
     {
         Workspace = workspace;
-        DisplayKeys = new ObservableCollection<DisplayKey>(
-            Workspace.Paths
-                .Select(path => new DisplayKey(Workspace)
-                {
-                    IsTranslated = Workspace.GetText(path) != null, 
-                    Key = path
-                })
-            );
-        (AllNodes, DisplayNodes) = AssembleDisplayNodes();
-        _currentKey = Workspace.Paths.Count > 0 ? Workspace.Paths[0] : "";
-
-        var paths = Workspace.Paths;
-        for (int i = 0; i < paths.Count; i++)
-        {
-            var category = GetKeyCategory(paths[i]);
-            if (CategoryStarts.FindIndex(x => x.Category == category) == -1)
-                CategoryStarts.Add((category, i));
-        }
         
-        CascadeCheckTranslatedNodes(DisplayNodes);
-    }
-
-    private (ObservableCollection<DisplayNode>, ObservableCollection<DisplayNode>) AssembleDisplayNodes()
-    {
-        var allNodes = new ObservableCollection<DisplayNode>();
-        var nodes = new ObservableCollection<DisplayNode>();
-
-        foreach (var path in Workspace.Paths)
-        {
-            ObservableCollection<DisplayNode> nextList = nodes;
-            DisplayNode? targetNode = null;
-            DisplayNode? parentNode = null;
-
-            var parts = Constants.TurnToDisplayPath(Workspace, path);
-
-            for (int i = 0; i < parts.Length; i++)
-            {
-                targetNode = nextList.FirstOrDefault(x => x.Label == parts[i]);
-
-                if (targetNode == null)
-                {
-                    nextList.Add(targetNode = new DisplayNode(parentNode)
-                    {
-                        Label = parts[i]
-                    });
-                    allNodes.Add(targetNode);
-                }
-
-                nextList = targetNode.SubNodes;
-                parentNode = targetNode;
-            }
-
-            if (targetNode != null)
-                targetNode.TerminalKey = path;
-        }
-
-        return (allNodes, nodes);
-    }
-
-    private bool CascadeCheckTranslatedNodes(ObservableCollection<DisplayNode> nodes)
-    {
-        bool allTranslated = true;
+        _englishJsonNavData = NavigationData.Create(workspace, false);
+        _alphabeticNavData = NavigationData.Create(workspace, true);
+        NavData = _englishJsonNavData;
         
-        foreach (var node in nodes)
-        {
-            bool translated = node.TerminalKey != null
-                ? Workspace.GetText(node.TerminalKey) != null
-                : CascadeCheckTranslatedNodes(node.SubNodes);
-            
-            node.AllItemsTranslated = translated;
-            allTranslated = allTranslated && translated;
-        }
-
-        return allTranslated;
+        // This length check is *in 99% of cases* not needed
+        _currentKey = NavData.ListView.Count > 0 ? NavData.ListView[0].FullPath : "";
+        
+        foreach (var node in NavData.ListView)
+            node.UpdateTranslationStatus(workspace);
     }
+
+    public List<int> AllowedIndents { get; } = [2, 3, 4];
+
+    public List<string> DisplayKeySortOptions { get; } = ["Like EN.json", "Alphabetically"];
 
     public Workspace Workspace { get; }
 
-    public ObservableCollection<DisplayKey> DisplayKeys { get; }
-    
-    public ObservableCollection<DisplayNode> AllNodes { get; }
-    public ObservableCollection<DisplayNode> DisplayNodes { get; }
+    [ObservableProperty]
+    private string _selectedDisplayKeySorting = "Like EN.json";
+
+    partial void OnSelectedDisplayKeySortingChanged(string value)
+    {
+        var currentKey = CurrentKey;
+        
+        switch (value)
+        {
+            case "Alphabetically":
+                NavData = _alphabeticNavData;
+                break;
+            default:
+                NavData = _englishJsonNavData;
+                break;
+        }
+        
+        foreach (var node in NavData.ListView)
+            node.UpdateTranslationStatus(Workspace);
+        
+        // Has to be done *after* setting the new item source!
+        var updatedNode = NavData.ListView.FirstOrDefault(x => x.FullPath == currentKey);
+        CurrentKeyIndex = updatedNode == null ? 0 : NavData.ListView.IndexOf(updatedNode);
+    }
+
+    [ObservableProperty]
+    private NavigationData _navData;
+
+    private NavigationData _alphabeticNavData;
+    private NavigationData _englishJsonNavData;
     
     [ObservableProperty]
     [NotifyPropertyChangedFor(nameof(PaginationText))]
@@ -144,30 +212,31 @@ public partial class TranslationWindowViewModel : ObservableObject
     [NotifyCanExecuteChangedFor(nameof(PreviousCategoryCommand))]
     [NotifyCanExecuteChangedFor(nameof(NextCategoryCommand))]
     private int _currentKeyIndex;
+    
+    partial void OnCurrentKeyIndexChanged(int value)
+    {
+        bool isValidIndex = 0 <= value && value < NavData.ListView.Count;
+        CurrentKey = isValidIndex ? NavData.ListView[CurrentKeyIndex].FullPath : "";
+        SelectedTreeViewItem = isValidIndex ? NavData.ListView[CurrentKeyIndex] : null;
+    }
 
     [ObservableProperty]
     private object? _selectedTreeViewItem;
-
+    
     partial void OnSelectedTreeViewItemChanged(object? value)
     {
-        if (value is not DisplayNode node || node.TerminalKey == null)
+        if (value is not DisplayNode node || !node.IsTerminalNode)
             return;
 
-        var index = Workspace.Paths.IndexOf(node.TerminalKey);
+        var index = NavData.ListView.IndexOf(node);
 
         if (index == -1)
             return;
 
         CurrentKeyIndex = index;
     }
-    
-    partial void OnCurrentKeyIndexChanged(int value)
-    {
-        bool isValidIndex = 0 <= value && value < Workspace.Paths.Count;
-        CurrentKey = isValidIndex ? Workspace.Paths[CurrentKeyIndex] : "";
-    }
 
-    private string GetKeyCategory(string key)
+    private static string GetKeyCategory(string key)
     {
         var category = key.Split(':')[0];
         return category.Length > 1 ? char.ToUpper(category[0]) + category[1..] : category;
@@ -197,7 +266,7 @@ public partial class TranslationWindowViewModel : ObservableObject
     [NotifyPropertyChangedFor(nameof(CurrentKeyTranslation))]
     [NotifyPropertyChangedFor(nameof(CurrentKeyEnglishDisplay))]
     [NotifyPropertyChangedFor(nameof(CurrentKeyTranslationDisplay))]
-    private string _currentKey;
+    private string _currentKey = "";
 
     public string CurrentKeyNotes => Constants.GetTranslationNotes(Workspace, CurrentKey);
 
@@ -212,19 +281,10 @@ public partial class TranslationWindowViewModel : ObservableObject
             if (SetProperty(ref oldValue, value) && value != null)
             {
                 Workspace.SetText(CurrentKey, value);
-                DisplayKeys[CurrentKeyIndex].IsTranslated = true;
+                NavData.ListView[CurrentKeyIndex].UpdateTranslationStatus(Workspace);
+                
                 OnPropertyChanged(nameof(CurrentKeyTranslationDisplay));
                 OnPropertyChanged(nameof(TranslationProgress));
-
-                var displayNode = AllNodes.FirstOrDefault(x => x.TerminalKey == CurrentKey);
-
-                if (displayNode != null)
-                {
-                    displayNode.AllItemsTranslated = true;
-
-                    // TODO: Optimize this
-                    CascadeCheckTranslatedNodes(DisplayNodes);
-                }
             }
         }
     }
@@ -292,29 +352,29 @@ public partial class TranslationWindowViewModel : ObservableObject
     [RelayCommand(CanExecute = nameof(CanExecuteNextMainKey))]
     public void NextMainKey()
     {
-        CurrentKeyIndex = Math.Min(CurrentKeyIndex + 1, Workspace.Paths.Count - 1);
+        CurrentKeyIndex = Math.Min(CurrentKeyIndex + 1, NavData.ListView.Count - 1);
     }
 
-    public bool CanExecuteNextMainKey() => CurrentKeyIndex < Workspace.Paths.Count - 1;
+    public bool CanExecuteNextMainKey() => CurrentKeyIndex < NavData.ListView.Count - 1;
 
     [RelayCommand(CanExecute = nameof(CanExecutePreviousCategory))]
     public void PreviousCategory()
     {
         var category = GetKeyCategory(CurrentKey);
 
-        var currentIndex = CategoryStarts.FindIndex(x => x.Category == category);
+        var currentIndex = NavData.CategoryStarts.FindIndex(x => x.Category == category);
 
         if (currentIndex <= 0)
             return;
 
-        CurrentKeyIndex = CategoryStarts[currentIndex - 1].Start;
+        CurrentKeyIndex = NavData.CategoryStarts[currentIndex - 1].Start;
     }
 
     public bool CanExecutePreviousCategory()
     {
         var category = GetKeyCategory(CurrentKey);
 
-        var currentIndex = CategoryStarts.FindIndex(x => x.Category == category);
+        var currentIndex = NavData.CategoryStarts.FindIndex(x => x.Category == category);
 
         return currentIndex > 0;
     }
@@ -324,33 +384,34 @@ public partial class TranslationWindowViewModel : ObservableObject
     {
         var category = GetKeyCategory(CurrentKey);
 
-        var currentIndex = CategoryStarts.FindIndex(x => x.Category == category);
+        var currentIndex = NavData.CategoryStarts.FindIndex(x => x.Category == category);
 
-        if (currentIndex == -1 || currentIndex >= CategoryStarts.Count - 1)
+        if (currentIndex == -1 || currentIndex >= NavData.CategoryStarts.Count - 1)
             return;
 
-        CurrentKeyIndex = CategoryStarts[currentIndex + 1].Start;
+        CurrentKeyIndex = NavData.CategoryStarts[currentIndex + 1].Start;
     }
 
     public bool CanExecuteNextCategory()
     {
         var category = GetKeyCategory(CurrentKey);
 
-        var currentIndex = CategoryStarts.FindIndex(x => x.Category == category);
+        var currentIndex = NavData.CategoryStarts.FindIndex(x => x.Category == category);
 
-        return currentIndex != -1 && currentIndex < CategoryStarts.Count - 1;
+        return currentIndex != -1 && currentIndex < NavData.CategoryStarts.Count - 1;
     }
     
     [RelayCommand]
     public void NextUntranslatedKey(Control control)
     {
-        var keys = Workspace.Paths;
+        var keys = NavData.ListView;
 
-        var firstUntranslated = keys.FindIndex(x => Workspace.GetText(x) == null);
+        var firstUntranslated = keys.FirstOrDefault(x => Workspace.GetText(x.FullPath) == null);
+        var firstUntranslatedIndex = firstUntranslated != null ? keys.IndexOf(firstUntranslated) : -1;
 
-        if (firstUntranslated == -1)
-            firstUntranslated = keys.Count - 1; // TODO: Replace with a notification that no such keys exist
+        if (firstUntranslatedIndex == -1)
+            firstUntranslatedIndex = keys.Count - 1; // TODO: Replace with a notification that no such keys exist
 
-        CurrentKeyIndex = Math.Clamp(firstUntranslated, 0, keys.Count - 1);
+        CurrentKeyIndex = Math.Clamp(firstUntranslatedIndex, 0, keys.Count - 1);
     }
 }